refactor: Move ChartContainer to a new index file for cleaner imports
