+++ src/platform/src/common/features/airQuality-charts/ChartContainer/index.jsx
